+++ crates/swc_fast_ts_strip/tests/tsc/constructorImplementationWithDefaultValues.strip.js
@@ -1,0 +1,20 @@
+class C {
+    constructor(x);
+    constructor(x = 1) {
+        var y = x;
+    }
+}
+
+class D    {
+    constructor(x);
+    constructor(x   = null) {
+        var y = x;
+    }
+}
+
+class E                 {
+    constructor(x);
+    constructor(x    = null) {
+        var y = x;
+    }
+}